refactor: 도메인 패키지내 클래스 재분류
