+++ src/main/java/com/qbang/stockpedia/dto/EmfFactory.java
@@ -1,4 +1,4 @@
-package com.qbang.stockpedia.domain;
+package com.qbang.stockpedia.dto;
 
 import org.springframework.stereotype.Component;
