+++ src/main/java/com/qbang/stockpedia/dto/TestVO.java
@@ -1,4 +1,4 @@
-package com.qbang.stockpedia.domain;
+package com.qbang.stockpedia.dto;
 
 public class TestVO {
     private String data;
